TR updates, first round
