+++ python-callable-instances/factorial.py
@@ -3,8 +3,6 @@ class Factorial:
         self.cache = {0: 1, 1: 1}
 
     def __call__(self, number):
-        if number in self.cache:
-            return self.cache[number]
-        else:
-            self.cache[number] = number * self.__call__(number - 1)
-            return self.cache[number]
+        if number not in self.cache:
+            self.cache[number] = number * self(number - 1)
+        return self.cache[number]
